fedmed_model/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
flwr>=1.7.0
opacus>=1.4.0
numpy>=1.26.3
pandas>=2.2.0
Pillow>=10.2.0
scikit-learn>=1.4.0
matplotlib>=3.8.2
seaborn>=0.13.2
pydicom>=2.4.4
opencv-python>=4.9.0.80
albumentations>=1.4.0
tqdm>=4.66.1
tensorboard>=2.15.2
fastapi>=0.109.2
uvicorn>=0.27.1
python-multipart>=0.0.9
pydantic>=2.6.1
